backend/database/graphs_cmd/requirements.txt
matplotlib==3.10.6
networkx==3.5
apache-age-python==0.0.7